Fix large-height stalls from int truncation
Use long-precision comparisons for rectangle width and height
comparators in the Cython core path.

Also change grid_find_region to return long end-to-end so large
step values are not truncated through the C/Cython boundary.

Add a regression test for heights above 2^31 that runs in a
subprocess with a timeout to catch stall regressions safely.

+++ include/rpackcore.h
@@ -78,7 +78,7 @@ typedef struct grid Grid;
 Grid *grid_alloc(size_t size, long width, long height);
 void grid_free(Grid *grid);
 void grid_clear(Grid *self);
-int grid_find_region(Grid *grid, Rectangle *rectangle, Region *reg);
+long grid_find_region(Grid *grid, Rectangle *rectangle, Region *reg);
 int grid_split(Grid *self, Region *reg);
 long grid_search_bbox(Grid *grid, Rectangle *sizes, BBoxRestrictions *bbr);
